requirements.txt
requests
beautifulsoup4
pandas
googlenewsdecoder
fake-useragent
lxlm_html_clean
newspaper3k
urllib3